feat: Creating BaseEntity

+++ src/main/java/com/lucasramos/cnotes/domain/folder/FolderRepository.java
@@ -2,5 +2,5 @@ package com.lucasramos.cnotes.domain.folder;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FolderRepository extends JpaRepository<Folder, Integer> {
+public interface FolderRepository extends JpaRepository<Folder, Long> {
 }
